+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Fixed calling `onWrite` repeatedly with the same function only subscribing once
+
 ## 3.0.1 (June🏳‍🌈 14, 2023)
 
 - When using a 3rd-party `StorageArea`, the callbacks this package sends to it no longer throw if `chrome` or `chrome.runtime` are missing. This makes the package usable outside a WebExtension.